+++ src/core/main.c
@@ -76,20 +76,6 @@ static struct PyModuleDef core_module_def = {
   .m_size = -1,
 };
 
-PyObject* init_dict;
-
-/**
- * The C code for runPythonSimple. The definition of runPythonSimple is in
- * `pyodide.js` for greater visibility.
- */
-int
-run_python_simple_inner(char* code)
-{
-  PyObject* result = PyRun_String(code, Py_file_input, init_dict, init_dict);
-  Py_XDECREF(result);
-  return result ? 0 : -1;
-}
-
 // from numpy_patch.c (no need for a header just for this)
 int
 numpy_patch_init();
@@ -103,11 +89,13 @@ get_python_stack_depth()
 
 /**
  * Bootstrap steps here:
- *  1. Initialize init_dict so that runPythonSimple will work.
+ *  1. Import _pyodide package (we depend on this in _pyodide_core)
  *  2. Initialize the different ffi components and create the _pyodide_core
  *     module
- *  3. Create a PyProxy wrapper around init_dict so that JavaScript can retreive
- *     PyProxies from the runPythonSimple namespace.
+ *  3. Create a PyProxy wrapper around _pyodide package so that JavaScript can
+ *     call into _pyodide._base.eval_code and
+ *     _pyodide._import_hook.register_js_finder (this happens in loadPyodide in
+ *     pyodide.js)
  */
 int
 main(int argc, char** argv)
@@ -115,14 +103,6 @@ main(int argc, char** argv)
   // This exits and prints a message to stderr on failure,
   // no status code to check.
   initialize_python();
-
-  // Once we initialize init_dict, runPythonSimple can work. This gives us a way
-  // to run Python code that works even if the rest of the initialization fails
-  // pretty badly.
-  init_dict = PyDict_New();
-  if (init_dict == NULL) {
-    FATAL_ERROR("Failed to create init_dict.");
-  }
 
   if (alignof(JsRef) != alignof(int)) {
     FATAL_ERROR("JsRef doesn't have the same alignment as int.");
@@ -133,7 +113,7 @@ main(int argc, char** argv)
 
   PyObject* _pyodide = NULL;
   PyObject* core_module = NULL;
-  JsRef init_dict_proxy = NULL;
+  JsRef _pyodide_proxy = NULL;
 
   _pyodide = PyImport_ImportModule("_pyodide");
   if (_pyodide == NULL) {
@@ -167,16 +147,15 @@ main(int argc, char** argv)
     FATAL_ERROR("Failed to add '_pyodide_core' module to modules dict.");
   }
 
-  // Enable JavaScript access to the globals from runPythonSimple.
-  init_dict_proxy = python2js(init_dict);
-  if (init_dict_proxy == NULL) {
-    FATAL_ERROR("Failed to create init_dict proxy.");
+  // Enable JavaScript access to the _pyodide module.
+  _pyodide_proxy = python2js(_pyodide);
+  if (_pyodide_proxy == NULL) {
+    FATAL_ERROR("Failed to create _pyodide proxy.");
   }
-  EM_ASM({ Module.init_dict = Module.hiwire.pop_value($0); }, init_dict_proxy);
+  EM_ASM({ Module._pyodide = Module.hiwire.pop_value($0); }, _pyodide_proxy);
 
   Py_CLEAR(_pyodide);
   Py_CLEAR(core_module);
-  hiwire_CLEAR(init_dict_proxy);
   emscripten_exit_with_live_runtime();
   return 0;
 }