+++ src/tests/test_jsproxy.py
@@ -1015,6 +1015,39 @@ def test_buffer(selenium):
     )
 
 
+@run_in_pyodide
+def test_buffer_to_file():
+    from js import Uint8Array
+
+    a = Uint8Array.new(range(10))
+    from tempfile import TemporaryFile
+
+    with TemporaryFile() as f:
+        a.to_file(f)
+        f.seek(0)
+        assert f.read() == a.to_bytes()
+
+        b = b"abcdef"
+        f.write(b)
+        f.seek(-len(b), 1)
+        a.from_file(f)
+        assert list(a.subarray(0, len(b)).to_bytes()) == list(b)
+
+
+@run_in_pyodide
+def test_buffer_into_file():
+    from js import Uint8Array
+
+    a = Uint8Array.new(range(10))
+    from tempfile import TemporaryFile
+
+    with TemporaryFile() as f:
+        b = a.to_bytes()
+        a._into_file(f)
+        f.seek(0)
+        assert f.read() == b
+
+
 def test_buffer_assign_back(selenium):
     result = selenium.run_js(
         """