+++ src/tests/test_pyodide.py
@@ -770,6 +770,31 @@ def test_reentrant_error(selenium):
     assert caught
 
 
+def test_reentrant_fatal(selenium_standalone):
+    selenium = selenium_standalone
+    assert selenium.run_js(
+        """
+        function f(){
+            pyodide.globals.get("trigger_fatal_error")();
+        }
+        self.success = true;
+        try {
+            pyodide.runPython(`
+                from _pyodide_core import trigger_fatal_error
+                from js import f
+                try:
+                    f()
+                except Exception as e:
+                    # This code shouldn't be executed
+                    import js
+                    js.success = False
+            `);
+        } catch(e){}
+        return success;
+        """
+    )
+
+
 def test_restore_error(selenium):
     # See PR #1816.
     selenium.run_js(
@@ -873,3 +898,33 @@ def test_custom_stdin_stdout(selenium_standalone_noload):
     )
     assert stdoutstrings == ["Python initialization complete", "something to stdout"]
     assert stderrstrings == ["something to stderr"]
+
+
+def test_home_directory(selenium_standalone_noload):
+    selenium = selenium_standalone_noload
+    home = "/home/custom_home"
+    selenium.run_js(
+        """
+        let pyodide = await loadPyodide({
+            indexURL : './',
+            homedir : "%s",
+        });
+        return pyodide.runPython(`
+            import os
+            os.getcwd() == "%s"
+        `)
+        """
+        % (home, home)
+    )
+
+
+def test_sys_path0(selenium):
+    selenium.run_js(
+        """
+        pyodide.runPython(`
+            import sys
+            import os
+            assert os.getcwd() == sys.path[0]
+        `)
+        """
+    )